Refactored the wheels logic to a new file
